update backend service

+++ Backend/Web.AppCore/Services/Order/OrderService.cs
@@ -9,6 +9,7 @@ using Web.AppCore.Interfaces.Repository;
 using Web.AppCore.Interfaces.Services;
 using Web.AppCore.Interfaces.Services.MessageQueue;
 using Web.Models.Entities;
+using Web.Models.Request;
 
 namespace Web.AppCore.Services
 {
@@ -16,10 +17,14 @@ namespace Web.AppCore.Services
     {
         private readonly IPublisherQueue _publisherQueue;
         private readonly IOrderUoW _orderUoW;
+        private readonly IOrderItemUoW _orderItemUoW;
+
+
         public OrderService(IPublisherQueue publisherQueue, IServiceProvider serviceProvider) : base(serviceProvider)
         {
             _publisherQueue = publisherQueue;
             _orderUoW = serviceProvider.GetRequiredService<IOrderUoW>();
+            _orderItemUoW = serviceProvider.GetRequiredService<IOrderItemUoW>();
         }
 
         /// <summary>
@@ -29,7 +34,20 @@ namespace Web.AppCore.Services
         /// <returns></returns>
         public async Task<bool> DeleteOrderAsync(string orderId)
         {
-            return await _orderUoW.Orders.DeleteOneAsync(orderId);
+            try
+            {
+                if (string.IsNullOrEmpty(orderId)) return false;
+                //Xóa chi tiết đơn hàng
+                var orderItems = await _orderItemUoW.OrderItems.GetAllAsync(x => x.order_id == orderId);
+                await _orderItemUoW.OrderItems.DeleteManyAsync(orderItems);
+                //Xóa đơn hàng
+                var resDelete = await _orderUoW.Orders.DeleteOneAsync(orderId);
+                return resDelete;
+            }
+            catch (Exception ex)
+            {
+                return false;
+            }
         }
 
         /// <summary>
@@ -37,9 +55,24 @@ namespace Web.AppCore.Services
         /// </summary>
         /// <param name="order"></param>
         /// <returns></returns>
-        public async Task<bool> InsertOrderAsync(Order order)
+        public async Task<bool> InsertOrderAsync(OrderRequest orderRequest)
         {
-            return await _orderUoW.Orders.InsertOneAsync(order);
+            try
+            {
+                var orderInsert = await _orderUoW.Orders.InsertOneAsync(orderRequest);
+                if (orderInsert == null) return false;
+
+                //Thêm chi tiết đơn hàng
+                if (orderRequest.order_items != null && orderRequest.order_items.Count > 0)
+                {
+                    await _orderItemUoW.OrderItems.InsertManyAsync(orderRequest.order_items);
+                }
+                return true;
+            }
+            catch (Exception ex)
+            {
+                return false;
+            }
         }
 
         /// <summary>
@@ -51,7 +84,7 @@ namespace Web.AppCore.Services
         {
             return await _orderUoW.Orders.UpdateOneAsync(order);
         }
-            
+
         public async Task<bool> UpdateOrderAsync()
         {
             return await _publisherQueue.PublishUpdateOrderAsync("Hello World");
